+++ src/handlers/index.ts
@@ -1,14 +1,3 @@
-export * from "./tsonBigint.js";
-export * from "./tsonDate.js";
-export * from "./tsonRegExp.js";
-export * from "./tsonSet.js";
-export * from "./tsonMap.js";
-export * from "./tsonUndefined.js";
-export * from "./tsonUnknownObjectGuard.js";
-export * from "./tsonNumberGuard.js";
-export * from "./tsonURL.js";
-export * from "./tsonSymbol.js";
-
 // Async types
 export * from "./tsonPromise.js";
 export * from "./tsonAsyncIterable.js";
