+++ src/index.ts
@@ -1,12 +1,13 @@
-// async
-export { createTsonAsync } from "./async/createTsonAsync.js";
-export { createTsonParseAsync } from "./async/deserializeAsync.js";
-export { createTsonStringifyAsync } from "./async/serializeAsync.js";
-
 // sync
 export { createTson } from "./sync/createTson.js";
 export { createTsonDeserialize, createTsonParser } from "./sync/deserialize.js";
 export { createTsonSerialize, createTsonStringify } from "./sync/serialize.js";
+export * from "./sync/handlers/index.js";
+
+// async
+export { createTsonAsync } from "./async/createTsonAsync.js";
+export { createTsonParseAsync } from "./async/deserializeAsync.js";
+export { createTsonStringifyAsync } from "./async/serializeAsync.js";
 export * from "./handlers/index.js";
 
 // types
