+++ src/sync/handlers/tsonSymbol.ts
@@ -1,4 +1,4 @@
-import { TsonType } from "../sync/syncTypes.js";
+import { TsonType } from "../syncTypes.js";
 
 export const tsonSymbol = <T extends symbol>(
 	symbol: T,
